:hammer: add header and sections, fix date & email bug. Create function complete.

+++ database/migrations/2022_12_02_211218_create_empresa_table.php
@@ -17,10 +17,10 @@ return new class extends Migration
             $table->id();
             $table->timestamps();
             $table->string('nombre', 120);
-            $table->tinyInteger('cuit');
+            $table->bigInteger('cuit');
             $table->string('direccion');
             $table->string('email',120)->nullable();
-            $table->tinyInteger('telefono')->nullable();
+            $table->bigInteger('telefono')->nullable();
         });
     }
